requirements.txt

# Database
pymongo==4.6.1
bson==0.5.10

# Environment and time

pytz==2024.1

# AWS
boto3==1.34.81
botocore==1.34.81

# Core dependencies
langchain>=0.1.0
langchain-community>=0.0.27
langchain-core>=0.1.0
langchain-openai>=0.0.8
langchain-text-splitters>=0.0.1

# OpenAI and embeddings
openai>=1.12.0

# Vector store
faiss-cpu>=1.7.4
chromadb==0.4.22

# PDF processing
PyPDF2>=3.0.1
pypdf==3.17.1

# Environment
python-dotenv>=1.0.0

# Utilities
pydantic>=2.6.1
tiktoken>=0.6.0

# Optional but recommended for better performance
numpy>=1.24.4
pandas>=2.0.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.2
torch>=2.0.0

# Web framework (if needed)
fastapi>=0.110.0
httpx>=0.27.0
flask>=2.3.3
Werkzeug>=2.3.7
flask-cors>=3.0.10

# Additional utilities
typing-extensions>=4.9.0
sqlalchemy>=2.0.27

#pdf generation
reportlab==4.0.0

